big updates

+++ bashconfig/.config/bashrc/aliases.bashrc
@@ -532,37 +532,42 @@ r_extract() {
 r_vidToGif() {
 	# Usage: r_vidToGif -i vid.mp4 --- -o file.gif
 
-	delimiter='---'
+	if command -v gifski >/dev/null; then
+		delimiter='---'
 
-	trap 'exit 2' SIGINT
+		trap 'exit 2' SIGINT
 
-	tmp="$(mktemp -td gifski-video.XXXXXX)"
+		tmp="$(mktemp -td gifski-video.XXXXXX)"
 
-	trap 'rm -rf "$tmp"; exit 1' EXIT
-	trap 'rm -rf "$tmp"; exit 2' INT
+		trap 'rm -rf "$tmp"; exit 1' EXIT
+		trap 'rm -rf "$tmp"; exit 2' INT
 
-	chmod 700 "$tmp"
+		chmod 700 "$tmp"
 
-	ffmpeg_opts=()
-	gifski_opts=()
+		ffmpeg_opts=()
+		gifski_opts=()
 
-	is_delimiter_encountered=0
-	for opt in "$@"; do
-		if [[ "$opt" == "$delimiter" ]]; then
-			is_delimiter_encountered=1
-			continue
-		fi
+		is_delimiter_encountered=0
+		for opt in "$@"; do
+			if [[ "$opt" == "$delimiter" ]]; then
+				is_delimiter_encountered=1
+				continue
+			fi
 
-		if ((!is_delimiter_encountered)); then
-			ffmpeg_opts+=("$opt")
-		else
-			gifski_opts+=("$opt")
-		fi
-	done
+			if ((!is_delimiter_encountered)); then
+				ffmpeg_opts+=("$opt")
+			else
+				gifski_opts+=("$opt")
+			fi
+		done
 
-	ffmpeg "${ffmpeg_opts[@]}" -hide_banner "$tmp"/frame%04d.png
-	gifski "${gifski_opts[@]}" "$tmp"/frame*.png
-	echo # gifski doesn't add a newline
+		ffmpeg "${ffmpeg_opts[@]}" -hide_banner "$tmp"/frame%04d.png
+		gifski "${gifski_opts[@]}" "$tmp"/frame*.png
+		echo # gifski doesn't add a newline
+	else
+		echo "command: gifski not found"
+	fi
+
 }
 
 # run: run newsboat
